feat: Add util for creating maps with default values
This commit adds `createMap`, a utility function that creates a
javascript `Map` object from an object's entries; returning a function
that takes a key used to get a value. It accepts an optional default
value to return when the key is not in the map. The `createMap`
function is defined in its own file inside the `utils` folder,
where it is also exported along with other utility functions.

The new `createMap` is used in the `RegistrationWidget` component
to create a map of `content` object's properties that is passed to
the `renderFormField` function. The original version of the `createMap`
function used inside `RegistrationWidget` was removed and replaced
with a reference to the new file.